0.14 Release

+++ frigate/record/maintainer.py
@@ -3,12 +3,11 @@
 import asyncio
 import datetime
 import logging
-import multiprocessing as mp
 import os
-import queue
 import random
 import string
 import threading
+import time
 from collections import defaultdict
 from multiprocessing.synchronize import Event as MpEvent
 from pathlib import Path
@@ -17,6 +16,9 @@ from typing import Any, Optional, Tuple
 import numpy as np
 import psutil
 
+from frigate.comms.config_updater import ConfigSubscriber
+from frigate.comms.detections_updater import DetectionSubscriber, DetectionTypeEnum
+from frigate.comms.inter_process import InterProcessRequestor
 from frigate.config import FrigateConfig, RetainModeEnum
 from frigate.const import (
     CACHE_DIR,
@@ -27,8 +29,6 @@ from frigate.const import (
     RECORD_DIR,
 )
 from frigate.models import Event, Recordings
-from frigate.types import FeatureMetricsTypes
-from frigate.util.image import area
 from frigate.util.services import get_video_properties
 
 logger = logging.getLogger(__name__)
@@ -38,16 +38,21 @@ QUEUE_READ_TIMEOUT = 0.00001  # seconds
 
 class SegmentInfo:
     def __init__(
-        self, motion_box_count: int, active_object_count: int, average_dBFS: int
+        self,
+        motion_count: int,
+        active_object_count: int,
+        region_count: int,
+        average_dBFS: int,
     ) -> None:
-        self.motion_box_count = motion_box_count
+        self.motion_count = motion_count
         self.active_object_count = active_object_count
+        self.region_count = region_count
         self.average_dBFS = average_dBFS
 
     def should_discard_segment(self, retain_mode: RetainModeEnum) -> bool:
         return (
             retain_mode == RetainModeEnum.motion
-            and self.motion_box_count == 0
+            and self.motion_count == 0
             and self.average_dBFS == 0
         ) or (
             retain_mode == RetainModeEnum.active_objects
@@ -56,22 +61,16 @@ class SegmentInfo:
 
 
 class RecordingMaintainer(threading.Thread):
-    def __init__(
-        self,
-        config: FrigateConfig,
-        inter_process_queue: mp.Queue,
-        object_recordings_info_queue: mp.Queue,
-        audio_recordings_info_queue: Optional[mp.Queue],
-        process_info: dict[str, FeatureMetricsTypes],
-        stop_event: MpEvent,
-    ):
+    def __init__(self, config: FrigateConfig, stop_event: MpEvent):
         threading.Thread.__init__(self)
         self.name = "recording_maintainer"
         self.config = config
-        self.inter_process_queue = inter_process_queue
-        self.object_recordings_info_queue = object_recordings_info_queue
-        self.audio_recordings_info_queue = audio_recordings_info_queue
-        self.process_info = process_info
+
+        # create communication for retained recordings
+        self.requestor = InterProcessRequestor()
+        self.config_subscriber = ConfigSubscriber("config/record/")
+        self.detection_subscriber = DetectionSubscriber(DetectionTypeEnum.all)
+
         self.stop_event = stop_event
         self.object_recordings_info: dict[str, list] = defaultdict(list)
         self.audio_recordings_info: dict[str, list] = defaultdict(list)
@@ -83,7 +82,7 @@ class RecordingMaintainer(threading.Thread):
             for d in os.listdir(CACHE_DIR)
             if os.path.isfile(os.path.join(CACHE_DIR, d))
             and d.endswith(".mp4")
-            and not d.startswith("clip_")
+            and not d.startswith("preview_")
         ]
 
         files_in_use = []
@@ -166,6 +165,7 @@ class RecordingMaintainer(threading.Thread):
                 Event.select(
                     Event.start_time,
                     Event.end_time,
+                    Event.data,
                 )
                 .where(
                     Event.camera == camera,
@@ -183,12 +183,13 @@ class RecordingMaintainer(threading.Thread):
         recordings_to_insert: list[Optional[Recordings]] = await asyncio.gather(*tasks)
 
         # fire and forget recordings entries
-        self.inter_process_queue.put(
-            (INSERT_MANY_RECORDINGS, [r for r in recordings_to_insert if r is not None])
+        self.requestor.send_data(
+            INSERT_MANY_RECORDINGS,
+            [r for r in recordings_to_insert if r is not None],
         )
 
     async def validate_and_move_segment(
-        self, camera: str, events: Event, recording: dict[str, any]
+        self, camera: str, events: list[Event], recording: dict[str, any]
     ) -> None:
         cache_path = recording["cache_path"]
         start_time = recording["start_time"]
@@ -196,7 +197,7 @@ class RecordingMaintainer(threading.Thread):
         # Just delete files if recordings are turned off
         if (
             camera not in self.config.cameras
-            or not self.process_info[camera]["record_enabled"].value
+            or not self.config.cameras[camera].record.enabled
         ):
             Path(cache_path).unlink(missing_ok=True)
             self.end_time_cache.pop(cache_path, None)
@@ -256,6 +257,7 @@ class RecordingMaintainer(threading.Thread):
                     duration,
                     cache_path,
                     record_mode,
+                    event.data["type"] == "api",
                 )
             # if it doesn't overlap with an event, go ahead and drop the segment
             # if it ends more than the configured pre_capture for the camera
@@ -294,7 +296,9 @@ class RecordingMaintainer(threading.Thread):
     def segment_stats(
         self, camera: str, start_time: datetime.datetime, end_time: datetime.datetime
     ) -> SegmentInfo:
+        video_frame_count = 0
         active_count = 0
+        region_count = 0
         motion_count = 0
         for frame in self.object_recordings_info[camera]:
             # frame is after end time of segment
@@ -304,6 +308,7 @@ class RecordingMaintainer(threading.Thread):
             if frame[0] < start_time.timestamp():
                 continue
 
+            video_frame_count += 1
             active_count += len(
                 [
                     o
@@ -311,8 +316,8 @@ class RecordingMaintainer(threading.Thread):
                     if not o["false_positive"] and o["motionless_count"] == 0
                 ]
             )
-
-            motion_count += sum([area(box) for box in frame[2]])
+            motion_count += len(frame[2])
+            region_count += len(frame[3])
 
         audio_values = []
         for frame in self.audio_recordings_info[camera]:
@@ -332,7 +337,9 @@ class RecordingMaintainer(threading.Thread):
 
         average_dBFS = 0 if not audio_values else np.average(audio_values)
 
-        return SegmentInfo(motion_count, active_count, round(average_dBFS))
+        return SegmentInfo(
+            motion_count, active_count, region_count, round(average_dBFS)
+        )
 
     async def move_segment(
         self,
@@ -342,11 +349,12 @@ class RecordingMaintainer(threading.Thread):
         duration: float,
         cache_path: str,
         store_mode: RetainModeEnum,
+        manual_event: bool = False,  # if this segment is being moved due to a manual event
     ) -> Optional[Recordings]:
         segment_info = self.segment_stats(camera, start_time, end_time)
 
         # check if the segment shouldn't be stored
-        if segment_info.should_discard_segment(store_mode):
+        if not manual_event and segment_info.should_discard_segment(store_mode):
             Path(cache_path).unlink(missing_ok=True)
             self.end_time_cache.pop(cache_path, None)
             return
@@ -411,17 +419,19 @@ class RecordingMaintainer(threading.Thread):
                 )
 
                 return {
-                    Recordings.id: f"{start_time.timestamp()}-{rand_id}",
-                    Recordings.camera: camera,
-                    Recordings.path: file_path,
-                    Recordings.start_time: start_time.timestamp(),
-                    Recordings.end_time: end_time.timestamp(),
-                    Recordings.duration: duration,
-                    Recordings.motion: segment_info.motion_box_count,
+                    Recordings.id.name: f"{start_time.timestamp()}-{rand_id}",
+                    Recordings.camera.name: camera,
+                    Recordings.path.name: file_path,
+                    Recordings.start_time.name: start_time.timestamp(),
+                    Recordings.end_time.name: end_time.timestamp(),
+                    Recordings.duration.name: duration,
+                    Recordings.motion.name: segment_info.motion_count,
                     # TODO: update this to store list of active objects at some point
-                    Recordings.objects: segment_info.active_object_count,
-                    Recordings.dBFS: segment_info.average_dBFS,
-                    Recordings.segment_size: segment_size,
+                    Recordings.objects.name: segment_info.active_object_count
+                    + (1 if manual_event else 0),
+                    Recordings.regions.name: segment_info.region_count,
+                    Recordings.dBFS.name: segment_info.average_dBFS,
+                    Recordings.segment_size.name: segment_size,
                 }
         except Exception as e:
             logger.error(f"Unable to store recording segment {cache_path}")
@@ -433,30 +443,50 @@ class RecordingMaintainer(threading.Thread):
         return None
 
     def run(self) -> None:
-        camera_count = sum(camera.enabled for camera in self.config.cameras.values())
         # Check for new files every 5 seconds
         wait_time = 0.0
-        while not self.stop_event.wait(wait_time):
+        while not self.stop_event.is_set():
+            time.sleep(wait_time)
+
+            if self.stop_event.is_set():
+                break
+
             run_start = datetime.datetime.now().timestamp()
+
+            # check if there is an updated config
+            while True:
+                (
+                    updated_topic,
+                    updated_record_config,
+                ) = self.config_subscriber.check_for_update()
+
+                if not updated_topic:
+                    break
+
+                camera_name = updated_topic.rpartition("/")[-1]
+                self.config.cameras[camera_name].record = updated_record_config
+
             stale_frame_count = 0
             stale_frame_count_threshold = 10
             # empty the object recordings info queue
             while True:
-                try:
+                (topic, data) = self.detection_subscriber.get_data(
+                    timeout=QUEUE_READ_TIMEOUT
+                )
+
+                if not topic:
+                    break
+
+                if topic == DetectionTypeEnum.video:
                     (
                         camera,
                         frame_time,
                         current_tracked_objects,
                         motion_boxes,
                         regions,
-                    ) = self.object_recordings_info_queue.get(
-                        True, timeout=QUEUE_READ_TIMEOUT
-                    )
+                    ) = data
 
-                    if frame_time < run_start - stale_frame_count_threshold:
-                        stale_frame_count += 1
-
-                    if self.process_info[camera]["record_enabled"].value:
+                    if self.config.cameras[camera].record.enabled:
                         self.object_recordings_info[camera].append(
                             (
                                 frame_time,
@@ -465,55 +495,30 @@ class RecordingMaintainer(threading.Thread):
                                 regions,
                             )
                         )
-                except queue.Empty:
-                    q_size = self.object_recordings_info_queue.qsize()
-                    if q_size > camera_count:
-                        logger.debug(
-                            f"object_recordings_info loop queue not empty ({q_size})."
+                elif topic == DetectionTypeEnum.audio:
+                    (
+                        camera,
+                        frame_time,
+                        dBFS,
+                        audio_detections,
+                    ) = data
+
+                    if self.config.cameras[camera].record.enabled:
+                        self.audio_recordings_info[camera].append(
+                            (
+                                frame_time,
+                                dBFS,
+                                audio_detections,
+                            )
                         )
-                    break
+                elif topic == DetectionTypeEnum.api:
+                    continue
+
+                if frame_time < run_start - stale_frame_count_threshold:
+                    stale_frame_count += 1
 
             if stale_frame_count > 0:
                 logger.debug(f"Found {stale_frame_count} old frames.")
-
-            # empty the audio recordings info queue if audio is enabled
-            if self.audio_recordings_info_queue:
-                stale_frame_count = 0
-
-                while True:
-                    try:
-                        (
-                            camera,
-                            frame_time,
-                            dBFS,
-                            audio_detections,
-                        ) = self.audio_recordings_info_queue.get(
-                            True, timeout=QUEUE_READ_TIMEOUT
-                        )
-
-                        if frame_time < run_start - stale_frame_count_threshold:
-                            stale_frame_count += 1
-
-                        if self.process_info[camera]["record_enabled"].value:
-                            self.audio_recordings_info[camera].append(
-                                (
-                                    frame_time,
-                                    dBFS,
-                                    audio_detections,
-                                )
-                            )
-                    except queue.Empty:
-                        q_size = self.audio_recordings_info_queue.qsize()
-                        if q_size > camera_count:
-                            logger.debug(
-                                f"object_recordings_info loop audio queue not empty ({q_size})."
-                            )
-                        break
-
-                if stale_frame_count > 0:
-                    logger.error(
-                        f"Found {stale_frame_count} old audio frames, segments from recordings may be missing"
-                    )
 
             try:
                 asyncio.run(self.move_files())
@@ -525,4 +530,7 @@ class RecordingMaintainer(threading.Thread):
             duration = datetime.datetime.now().timestamp() - run_start
             wait_time = max(0, 5 - duration)
 
+        self.requestor.stop()
+        self.config_subscriber.stop()
+        self.detection_subscriber.stop()
         logger.info("Exiting recording maintenance...")
